Submit button is now adding thought to the database
